+++ SharpDPAPI/Commands/PS.cs
@@ -1,0 +1,57 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+
+namespace SharpDPAPI.Commands
+{
+    public class PS : ICommand
+    {
+        public static string CommandName => "ps";
+
+        public void Execute(Dictionary<string, string> arguments)
+        {
+            Console.WriteLine("\r\n[*] Action: Describe PSCredential .xml");
+
+            string target = "";
+            bool unprotect = false;         // whether to force CryptUnprotectData()
+
+            if (arguments.ContainsKey("/unprotect"))
+            {
+                Console.WriteLine("\r\n[*] Using CryptUnprotectData() for decryption.");
+                unprotect = true;
+            }
+            Console.WriteLine();
+
+            if (arguments.ContainsKey("/target"))
+            {
+                target = arguments["/target"];
+            }
+            else
+            {
+                Console.WriteLine("[X] A /target:<BASE64 | file.bin> must be supplied!");
+                return;
+            }
+
+            // {GUID}:SHA1 keys are the only ones that don't start with /
+            Dictionary<string, string> masterkeys = new Dictionary<string, string>();
+            foreach (KeyValuePair<string, string> entry in arguments)
+            {
+                if (!entry.Key.StartsWith("/"))
+                {
+                    masterkeys.Add(entry.Key, entry.Value);
+                }
+            }
+            if (arguments.ContainsKey("/pvk"))
+            {
+                // use a domain DPAPI backup key to triage masterkeys
+                masterkeys = SharpDPAPI.Dpapi.PVKTriage(arguments);
+            }
+            else if (arguments.ContainsKey("/mkfile"))
+            {
+                masterkeys = SharpDPAPI.Helpers.ParseMasterKeyFile(arguments["/mkfile"]);
+            }
+
+            Triage.TriagePSCredFile(masterkeys, target, unprotect);
+        }
+    }
+}