+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2019-07-25
+
+### Added
+* **ps** command to decrypt exported PSCredential xmls (thanks for the idea @gentilkiwi ;)
+* **blob** section for the README
+
+### Changed
+* **blob** command outputs hex if the data doesn't appear to be text
+
 
 ## [1.4.0] - 2019-05-22